trianed on all data, not much better

+++ src/models/model.py
@@ -40,27 +40,13 @@ class LightningBert(LightningModule):
         self.log("train-loss", loss)
         return loss
 
-    #def forward(self, tokenbatch):
-        #bertput = self.bert(tokenbatch[0], tokenbatch[1])
-        #output = self.class_layer(bertput.pooler_output)
-        #return output
-
-    #def training_step(self, batch, batch_idx):
-        #inputs, labels = batch
-        ## Forward pass
-        #outputs = self(inputs)
-        ## TODO: implement logging of the loss
-        #loss = self.loss(outputs, labels)
-        #self.log("train-loss", loss)
-        #return loss
-
     def configure_optimizers(self):
         return optim.Adam(self.parameters(), lr=self.lr)
     
     def train_dataloader(self):
         inputs = torch.load(self.cfg["train"]["datapaths"][0])
         labels = torch.load(self.cfg["train"]["datapaths"][1])
-        train = TensorDataset(inputs[0], inputs[1], labels)
+        train = TensorDataset(inputs[0].type(torch.int64), inputs[1].type(torch.int64), labels.float())
         train_loader = DataLoader(train, batch_size=self.batch_size, shuffle=True)
         return train_loader
 
@@ -68,7 +54,7 @@ class LightningBert(LightningModule):
         inputs = torch.load(self.cfg["train"]["datapaths"][2])
         labels = torch.load(self.cfg["train"]["datapaths"][3])
         #val = TensorDataset(inputs, labels)
-        val = TensorDataset(inputs[0], inputs[1], labels)
+        val = TensorDataset(inputs[0].type(torch.int64), inputs[1].type(torch.int64), labels.float())
         val_loader = DataLoader(val, batch_size=self.batch_size, shuffle=True)
         return val_loader
     
@@ -76,8 +62,3 @@ class LightningBert(LightningModule):
         inputs, mask, labels = batch
         outputs = self(inputs, mask)
         return {"val_loss": self.loss(outputs, labels)}
-
-    #def validation_step(self, batch, batch_idx):
-        #inputs, labels = batch
-        #outputs = self(inputs)
-        #return {"val_loss": self.loss(outputs, labels)}
